backend/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
langchain-groq>=0.2.0
langgraph>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0
